feat: implement check_perft (reads test EPDs)
bench results:
nodes: 45979680 nodes
time:  5476 milliseconds
nps:   8396177 nps

+++ src/rose/cmd/perft.cpp
@@ -10,11 +10,11 @@
 
 namespace rose::perft {
 
-  template <bool print> static auto core(const PrecompMoveGenInfo &movegen_precomp, const Position &position, usize depth) -> usize {
+  template <bool print> static auto core(const PrecompMoveGenInfo &movegen_precomp, const Position &position, usize depth) -> u64 {
     if (depth == 0)
       return 1;
 
-    usize result = 0;
+    u64 result = 0;
 
     MoveList moves;
     MoveGen movegen{position, movegen_precomp};
@@ -25,7 +25,7 @@ namespace rose::perft {
 
     for (Move m : moves) {
       const Position new_position = position.move(m);
-      const usize child = core<false>(movegen_precomp, new_position, depth - 1);
+      const u64 child = core<false>(movegen_precomp, new_position, depth - 1);
       if constexpr (print) {
         std::print("{}: {}\n", m, child);
       }
@@ -35,7 +35,7 @@ namespace rose::perft {
     return result;
   }
 
-  auto value(const Position &position, usize depth) -> usize {
+  auto value(const Position &position, usize depth) -> u64 {
     const PrecompMoveGenInfo movegen_precomp{position};
     return core<false>(movegen_precomp, position, depth);
   }
@@ -43,7 +43,7 @@ namespace rose::perft {
   auto run(const Position &position, usize depth) -> void {
     const auto start = time::Clock::now();
     const PrecompMoveGenInfo movegen_precomp{position};
-    const usize total = core<true>(movegen_precomp, position, depth);
+    const u64 total = core<true>(movegen_precomp, position, depth);
     const auto end = time::Clock::now();
 
     const time::FloatSeconds elapsed = end - start;
